requirements.txt
asgiref==3.5.2
cffi==1.15.1
colorama==0.4.6
crispy-bootstrap5==0.6
cssselect==1.2.0
Django==4.1.1
django-crispy-forms==1.14.0
django-filter==21.1
django-jsignature==0.11
django-safedelete==1.2.2
django-sequences==2.7
./utils/django-simple-history
django-sslserver==0.22
djangorestframework==3.13.1
importlib-metadata==4.11.3
lxml==4.9.2
Markdown==3.3.6
pdfkit==1.0.0
Pillow==9.0.1
psycopg2==2.9.5
pycparser==2.21
pycryptodome==3.16.0
pyquery==1.4.3
python-dotenv==0.19.2
python-json-logger==2.0.4
pytz==2022.1
qrcode==7.3.1
requests==2.28.0
sqlparse==0.4.2
tzdata==2021.5
Unidecode==1.3.6
zipp==3.7.0
pandas==1.4.4
django-pgtrigger==4.6.0
